requirements.txt
Django==3.2.15
django-import-export==2.5.0
django-mdeditor==0.1.18
Markdown==3.3.6
django-modeladmin-reorder==0.3.1
django-redis==5.0.0
django-simple-captcha==0.5.14
django-simpleui==2022.1
pymysql~=0.9.3
fast-captcha==0.1.3
requests==2.25.1
cryptography==37.0.4
wait-for-it==2.2.1
django-mptt==0.13.4
